added 'get_scale' function
The 'get_scale' function creates the ability to dynamically fetch
different scales by passing a string to the 'name' argument of the function representing the scale to return.

The scale is returned as a single array of integers.  Each integer
represents the interval (half-step) from the tonic.  The tonic is
represented as 0.

If the string passed to 'name' does not match an index in the
'scaleName' array, the function defaults to return an array with one
interval, the tonic.  This ensures that something will be played.

Opinion: This implementation of the function is convoluted,
probably due to the fact that I am still learning the fundamentals
of the ChucK language.  Perhaps one day this function gets a refactor.