added 'get_chord', 'play_chord' functions
Following the pattern established in 'get_scale' and 'play_scale',
we establish a convention for passing a chord name to the
'get_chord' functionthat returns an interval set.
We then pass a tonic freq and the interval set to the 'play_chord'
function, which passes each interval to a sine wave oscillator,
sets gain, and plays the chord for a period of one second.