added support for additional chords, array refactor
Added the diminished, and augmented triads, the suspended second and fourth, and added ninth (which doesn't quite sound right).  Adding
the ninth required implementing support for 4 and 5 note chords.

Also replaced the individual lines to choose scale and chords to play with
arrays and loops to play all the options.